0.2 Allow easier CSV reading options
For now these still need to be subclassed to be usable, but that's
better than overriding the main reading methods. (We can add a better
interface once it becomes more obvious what different ways comparisons
will need to override things.)

This also fixes a bug where the paths to the files were hardcoded,
ignoring whatever was passed in to the constructor, and renames
the 'official' source to solely be an 'external' one, as it's not
necessarily always going to be the official site.