0.3.0 #tds now also includes th cells
When fetching all the cells in a row, it usually doesn't really matter
whether they're actually a td or a th, and not having to care about
which is which makes things a lot simpler. This may break older scrapers
that _did_ have to care, so needs a next level version bump.